reafctor: move presentation logic to custom hooks
