+++ src/domain/entities/data-entry/amr-external/DatsetOptions.ts
@@ -1,0 +1,28 @@
+export const datasetOptions = [
+    {
+        label: "Dataset 1",
+        value: "DS1",
+    },
+    {
+        label: "Dataset 2",
+        value: "DS2",
+    },
+    {
+        label: "Dataset 3",
+        value: "DS3",
+    },
+    {
+        label: "Dataset 4",
+        value: "DS4",
+    },
+    {
+        label: "Dataset 5",
+        value: "DS5",
+    },
+    {
+        label: "Dataset 6",
+        value: "DS6",
+    },
+];
+
+export const UPLOADED_STATUS = "uploaded";
