+++ src/webapp/components/upload/UploadFiles.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useState } from "react";
 import {
     Backdrop,
     Button,
@@ -15,15 +15,14 @@ import { glassColors } from "../../pages/app/themes/dhis2.theme";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 import { UploadPrimaryFile } from "./UploadPrimaryFile";
 import { UploadSecondary } from "./UploadSecondary";
-import { useAppContext } from "../../contexts/app-context";
 import { useCurrentModuleContext } from "../../contexts/current-module-context";
-import { useCurrentOrgUnitContext } from "../../contexts/current-orgUnit-context";
-import { useCurrentPeriodContext } from "../../contexts/current-period-context";
-import { useSnackbar } from "@eyeseetea/d2-ui-components";
-import { Future } from "../../../domain/entities/Future";
 import { ImportSummary } from "../../../domain/entities/data-entry/ImportSummary";
 import { StyledLoaderContainer } from "./ConsistencyChecks";
 import { moduleProperties } from "../../../domain/utils/ModuleProperties";
+import { useBatchHandling } from "./hooks/useBatchHandling";
+import { BatchDropDown } from "./BatchDropDown";
+import { useDownloadEmptyTemplate } from "./hooks/useDownloadEmptyTemplate";
+import { useFileSubmission } from "./hooks/useFileSubmission";
 
 interface UploadFilesProps {
     changeStep: (step: number) => void;
@@ -37,35 +36,6 @@ interface UploadFilesProps {
     setSecondaryFileImportSummary: React.Dispatch<React.SetStateAction<ImportSummary | undefined>>;
 }
 
-const UPLOADED_STATUS = "uploaded";
-
-const datasetOptions = [
-    {
-        label: "Dataset 1",
-        value: "DS1",
-    },
-    {
-        label: "Dataset 2",
-        value: "DS2",
-    },
-    {
-        label: "Dataset 3",
-        value: "DS3",
-    },
-    {
-        label: "Dataset 4",
-        value: "DS4",
-    },
-    {
-        label: "Dataset 5",
-        value: "DS5",
-    },
-    {
-        label: "Dataset 6",
-        value: "DS6",
-    },
-];
-
 export const UploadFiles: React.FC<UploadFilesProps> = ({
     changeStep,
     primaryFile,
@@ -77,345 +47,32 @@ export const UploadFiles: React.FC<UploadFilesProps> = ({
     setPrimaryFileImportSummary,
     setSecondaryFileImportSummary,
 }) => {
-    const { compositionRoot } = useAppContext();
-    const snackbar = useSnackbar();
-    const [isValidated, setIsValidated] = useState(false);
-    const [isPrimaryFileValid, setIsPrimaryFileValid] = useState(false);
-    const [isSecondaryFileValid, setIsSecondaryFileValid] = useState(false);
-    const [previousUploadsBatchIds, setPreviousUploadsBatchIds] = useState<string[]>([]);
-    const [hasSecondaryFile, setHasSecondaryFile] = useState<boolean>(secondaryFile ? true : false);
-    const [importLoading, setImportLoading] = useState<boolean>(false);
-    const [previousBatchIdsLoading, setPreviousBatchIdsLoading] = useState<boolean>(true);
     const [loading, setLoading] = useState<boolean>(false);
     const {
         currentModuleAccess: { moduleName },
     } = useCurrentModuleContext();
-    const {
-        currentOrgUnitAccess: { orgUnitId, orgUnitName, orgUnitCode },
-    } = useCurrentOrgUnitContext();
-    const { currentPeriod } = useCurrentPeriodContext();
 
     const currentModuleProperties = moduleProperties.get(moduleName);
-    const [uploadFileType, setUploadFileType] = useState(
-        secondaryFile ? currentModuleProperties?.secondaryFileType : currentModuleProperties?.primaryFileType
+
+    const { changeFileType, continueClick, uploadFileType, setHasSecondaryFile, importLoading } = useFileSubmission(
+        changeStep,
+        primaryFile,
+        secondaryFile,
+        batchId,
+        setPrimaryFileImportSummary,
+        setSecondaryFileImportSummary
     );
 
-    useEffect(() => {
-        if (
-            currentModuleProperties?.isbatchReq ||
-            (currentModuleProperties?.isExternalSecondaryFile &&
-                uploadFileType === currentModuleProperties.secondaryFileType)
-        ) {
-            setPreviousBatchIdsLoading(true);
+    const {
+        previousBatchIdsLoading,
+        previousUploadsBatchIds,
+        changeBatchId,
+        setIsPrimaryFileValid,
+        setIsSecondaryFileValid,
+        isValidated,
+    } = useBatchHandling(uploadFileType, batchId, setBatchId, setLoading);
 
-            compositionRoot.glassUploads.getAMRUploadsForCurrentDataSubmission(orgUnitId, currentPeriod).run(
-                uploads => {
-                    const uniquePreviousBatchIds = _(
-                        uploads
-                            .filter(upload => upload.status.toLowerCase() !== UPLOADED_STATUS && upload.batchId !== "")
-                            .map(upload => upload.batchId)
-                    )
-                        .uniq()
-                        .value();
-                    setPreviousUploadsBatchIds(uniquePreviousBatchIds);
-                    const firstSelectableBatchId = datasetOptions.find(
-                        ({ value }) => !uniquePreviousBatchIds.includes(value)
-                    )?.value;
-                    setBatchId(firstSelectableBatchId || "");
-                    setPreviousBatchIdsLoading(false);
-                },
-                () => {
-                    snackbar.error(i18n.t("Error fetching previous uploads."));
-                    setPreviousBatchIdsLoading(false);
-                }
-            );
-        } else {
-            setPreviousBatchIdsLoading(false);
-        }
-    }, [
-        compositionRoot.glassUploads,
-        setBatchId,
-        snackbar,
-        moduleName,
-        currentPeriod,
-        orgUnitId,
-        currentModuleProperties,
-        uploadFileType,
-    ]);
-
-    useEffect(() => {
-        if (moduleProperties.get(moduleName)?.isbatchReq) {
-            if (batchId && isPrimaryFileValid) {
-                setIsValidated(true);
-            } else {
-                setIsValidated(false);
-            }
-        } else if (moduleProperties.get(moduleName)?.isSingleFileTypePerSubmission) {
-            if (isPrimaryFileValid || isSecondaryFileValid) setIsValidated(true);
-            else setIsValidated(false);
-        } else if (moduleProperties.get(moduleName)?.isExternalSecondaryFile) {
-            if (isPrimaryFileValid) setIsValidated(true);
-            else if (batchId && isSecondaryFileValid) setIsValidated(true);
-            else setIsValidated(false);
-        } else {
-            if (isPrimaryFileValid) setIsValidated(true);
-        }
-    }, [batchId, isPrimaryFileValid, isSecondaryFileValid, moduleName]);
-
-    const changeBatchId = async (event: React.ChangeEvent<{ value: unknown }>) => {
-        const batchId = event.target.value as string;
-        const primaryUploadId = localStorage.getItem("primaryUploadId");
-        const secondaryUploadId = localStorage.getItem("secondaryUploadId");
-        setBatchId(batchId);
-        if (primaryUploadId) {
-            setLoading(true);
-            compositionRoot.glassUploads.setBatchId({ id: primaryUploadId, batchId }).run(
-                () => {
-                    if (secondaryUploadId) {
-                        compositionRoot.glassUploads.setBatchId({ id: secondaryUploadId, batchId }).run(
-                            () => {
-                                setLoading(false);
-                            },
-                            () => {
-                                console.debug(`error occured while updating batch id to secondary upload in datastore`);
-                                setLoading(false);
-                            }
-                        );
-                    } else setLoading(false);
-                },
-                () => {
-                    console.debug(`error occured while updating batch id to primary upload in datastore`);
-                    setLoading(false);
-                }
-            );
-        } else if (secondaryUploadId) {
-            setLoading(true);
-            compositionRoot.glassUploads.setBatchId({ id: secondaryUploadId, batchId }).run(
-                () => {
-                    setLoading(false);
-                },
-                () => {
-                    console.debug(`error occured while updating batch id to secondary upload in datastore`);
-                    setLoading(false);
-                }
-            );
-        }
-    };
-
-    const changeFileType = (event: React.ChangeEvent<{ value: unknown }>) => {
-        const fileType = event.target.value as string;
-        setUploadFileType(fileType);
-    };
-
-    const uploadFileSubmissions = useCallback(() => {
-        if (primaryFile) {
-            setImportLoading(true);
-            Future.joinObj({
-                importPrimaryFileSummary: compositionRoot.fileSubmission.primaryFile(
-                    moduleName,
-                    primaryFile,
-                    batchId,
-                    currentPeriod,
-                    "CREATE_AND_UPDATE",
-                    orgUnitId,
-                    orgUnitName,
-                    orgUnitCode,
-                    true,
-                    ""
-                ),
-                importSecondaryFileSummary: secondaryFile
-                    ? compositionRoot.fileSubmission.secondaryFile(
-                          secondaryFile,
-                          batchId,
-                          moduleName,
-                          currentPeriod,
-                          "CREATE_AND_UPDATE",
-                          orgUnitId,
-                          orgUnitName,
-                          orgUnitCode,
-                          true,
-                          ""
-                      )
-                    : Future.success(undefined),
-            }).run(
-                ({ importPrimaryFileSummary, importSecondaryFileSummary }) => {
-                    setPrimaryFileImportSummary(importPrimaryFileSummary);
-
-                    if (importSecondaryFileSummary) {
-                        setSecondaryFileImportSummary(importSecondaryFileSummary);
-                    }
-
-                    const primaryUploadId = localStorage.getItem("primaryUploadId");
-
-                    if (primaryUploadId) {
-                        const secondaryUploadId = localStorage.getItem("secondaryUploadId");
-
-                        const params = secondaryUploadId
-                            ? {
-                                  primaryUploadId,
-                                  primaryImportSummaryErrors: {
-                                      nonBlockingErrors: importPrimaryFileSummary?.nonBlockingErrors || [],
-                                      blockingErrors: importPrimaryFileSummary?.blockingErrors || [],
-                                  },
-                                  secondaryUploadId,
-                                  secondaryImportSummaryErrors: {
-                                      nonBlockingErrors: importSecondaryFileSummary?.nonBlockingErrors || [],
-                                      blockingErrors: importSecondaryFileSummary?.blockingErrors || [],
-                                  },
-                              }
-                            : {
-                                  primaryUploadId,
-                                  primaryImportSummaryErrors: {
-                                      nonBlockingErrors: importPrimaryFileSummary?.nonBlockingErrors || [],
-                                      blockingErrors: importPrimaryFileSummary?.blockingErrors || [],
-                                  },
-                              };
-
-                        compositionRoot.glassUploads.saveImportSummaryErrorsOfFiles(params).run(
-                            () => {},
-                            () => {}
-                        );
-                    }
-
-                    setImportLoading(false);
-                    changeStep(2);
-                },
-                error => {
-                    setPrimaryFileImportSummary({
-                        status: "ERROR",
-                        importCount: { ignored: 0, imported: 0, deleted: 0, updated: 0 },
-                        nonBlockingErrors: [],
-                        blockingErrors: [{ error: error, count: 1 }],
-                    });
-                    setImportLoading(false);
-                    changeStep(2);
-                }
-            );
-        } //secondary file upload only
-        else if (secondaryFile) {
-            setImportLoading(true);
-            compositionRoot.fileSubmission
-                .secondaryFile(
-                    secondaryFile,
-                    batchId,
-                    moduleName,
-                    currentPeriod,
-                    "CREATE_AND_UPDATE",
-                    orgUnitId,
-                    orgUnitName,
-                    orgUnitCode,
-                    true,
-                    ""
-                )
-                .run(
-                    importSubstanceFileSummary => {
-                        if (importSubstanceFileSummary) {
-                            setPrimaryFileImportSummary(importSubstanceFileSummary);
-                        }
-                        setImportLoading(false);
-                        changeStep(2);
-                    },
-                    error => {
-                        setPrimaryFileImportSummary({
-                            status: "ERROR",
-                            importCount: { ignored: 0, imported: 0, deleted: 0, updated: 0 },
-                            nonBlockingErrors: [],
-                            blockingErrors: [{ error: error, count: 1 }],
-                        });
-                        setImportLoading(false);
-                        changeStep(2);
-                    }
-                );
-        }
-    }, [
-        primaryFile,
-        compositionRoot.fileSubmission,
-        compositionRoot.glassUploads,
-        moduleName,
-        batchId,
-        currentPeriod,
-        orgUnitId,
-        orgUnitName,
-        orgUnitCode,
-        secondaryFile,
-        setPrimaryFileImportSummary,
-        changeStep,
-        setSecondaryFileImportSummary,
-    ]);
-
-    const continueClick = () => {
-        if (!hasSecondaryFile) {
-            localStorage.removeItem("secondaryUploadId");
-            uploadFileSubmissions();
-        } else if (!moduleProperties.get(moduleName)?.isSecondaryRelated) {
-            uploadFileSubmissions();
-        }
-        //update the secondary file with primary file upload id.
-        else {
-            setImportLoading(true);
-            const primaryUploadId = localStorage.getItem("primaryUploadId");
-            const secondaryUploadId = localStorage.getItem("secondaryUploadId");
-            if (secondaryUploadId && primaryUploadId)
-                compositionRoot.glassDocuments.updateSecondaryFileWithPrimaryId(secondaryUploadId, primaryUploadId).run(
-                    () => {
-                        uploadFileSubmissions();
-                    },
-                    () => {
-                        console.debug("Error updating datastore");
-                    }
-                );
-        }
-    };
-
-    const downloadEmptyTemplate = useCallback(() => {
-        setLoading(true);
-
-        const fileType = uploadFileType === "Product Level Data" ? "PRODUCT" : "SUBSTANCE";
-        compositionRoot.fileSubmission.downloadEmptyTemplate(moduleName, fileType, orgUnitId).run(
-            file => {
-                //download file automatically
-                const downloadSimulateAnchor = document.createElement("a");
-                downloadSimulateAnchor.href = URL.createObjectURL(file);
-                const fileType = moduleProperties.get(moduleName)?.isSingleFileTypePerSubmission
-                    ? `-${uploadFileType}`
-                    : "";
-                downloadSimulateAnchor.download = `${moduleName}${fileType}-${orgUnitCode}-TEMPLATE.xlsx`;
-                // simulate link click
-                document.body.appendChild(downloadSimulateAnchor);
-                downloadSimulateAnchor.click();
-                setLoading(false);
-            },
-            (error: string) => {
-                snackbar.error("Error downloading file");
-                console.error(error);
-                setLoading(false);
-            }
-        );
-    }, [compositionRoot.fileSubmission, snackbar, orgUnitId, moduleName, orgUnitCode, uploadFileType]);
-
-    const getBatchIdDropDown = () => {
-        return (
-            <div className="batch-id">
-                <h3>{i18n.t("Batch ID")}</h3>
-                <FormControl variant="outlined" style={{ minWidth: 180 }}>
-                    <InputLabel id="dataset-label">{i18n.t("Choose a Dataset")}</InputLabel>
-                    <Select
-                        value={batchId}
-                        onChange={changeBatchId}
-                        label={i18n.t("Choose a Dataset")}
-                        labelId="dataset-label"
-                        MenuProps={{ disableScrollLock: true }}
-                    >
-                        {datasetOptions.map(({ label, value }) => (
-                            <MenuItem key={value} value={value} disabled={previousUploadsBatchIds.includes(value)}>
-                                {i18n.t(label)}
-                            </MenuItem>
-                        ))}
-                    </Select>
-                </FormControl>
-            </div>
-        );
-    };
+    const { downloadEmptyTemplate } = useDownloadEmptyTemplate(uploadFileType, setLoading);
 
     return (
         <ContentWrapper>
@@ -490,8 +147,13 @@ export const UploadFiles: React.FC<UploadFilesProps> = ({
                         <CircularProgress size={25} />
                     ) : (
                         currentModuleProperties.isExternalSecondaryFile &&
-                        uploadFileType === currentModuleProperties.secondaryFileType &&
-                        getBatchIdDropDown()
+                        uploadFileType === currentModuleProperties.secondaryFileType && (
+                            <BatchDropDown
+                                batchId={batchId}
+                                previousUploadsBatchIds={previousUploadsBatchIds}
+                                changeBatchId={changeBatchId}
+                            />
+                        )
                     )}
                 </>
             ) : (
@@ -517,7 +179,13 @@ export const UploadFiles: React.FC<UploadFilesProps> = ({
                                     />
                                 )}
                             </div>
-                            {moduleProperties.get(moduleName)?.isbatchReq && getBatchIdDropDown()}
+                            {moduleProperties.get(moduleName)?.isbatchReq && (
+                                <BatchDropDown
+                                    batchId={batchId}
+                                    previousUploadsBatchIds={previousUploadsBatchIds}
+                                    changeBatchId={changeBatchId}
+                                />
+                            )}
                         </>
                     )}
                 </>
